bot/requirements.txt
#requirements

aiogram
python-dotenv
aiosqlite
replicate
yookassa
python-dotenv>=1.0.0
aiohttp>=3.9.0
# YooKassa интеграция
yookassa>=3.0.0
